+++ test/io/jesi/backpack/clojure_test.clj
@@ -2,7 +2,7 @@
   (:refer-clojure :exclude [=])
   (:require
     [io.jesi.backpack :as bp]
-    [io.jesi.backpack.test.strict :refer :all])
+    [io.jesi.customs.strict :refer :all])
   (:import
     (java.net URI)
     (java.util HashMap)))
@@ -29,7 +29,7 @@
 (deftest defkw-type-test
 
   (testing "Registers keyword as type"
-    ;FIXME fails in test refresh (but not the first time it runs)
+    ;TODO fails in test refresh (but not the first time it runs)
     (is (instance? TestType tested))
     (is (instance? TestType and-worked))
     (is (instance? TestType lots-of-yelling)))